fix: script.module.tknorris.shared v0.0.3x
