--- tknorris.shared/.gitignore
@@ -1,4 +1,0 @@
-/.project
-/.pydevproject
-*.pyc
-*.pyo
