--- zips/tknorris.shared/changelog-0.0.3x.txt
@@ -1,11 +1,0 @@
-[B]Version 0.0.3x[/B]
-add countdown dialog
-
-[B]Version 0.0.3[/B]
-Force strptime import to avoid python bug
-
-[B]Version 0.0.2[/B]
-- Add optional component specific logging
-
-[B]Version 0.0.1[/B]
-- Initial Release.
